apis/core_app/pyproject.toml
[tool.poetry]
name = "core_app"
version = "0.1.0"
description = ""
authors = ["Renan Santana <renan_ota@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
psycopg2-binary = "^2.9.2"
uvicorn = "^0.15.0"
djangorestframework = "^3.12.4"
prefect = "^0.15.9"

[tool.poetry.dev-dependencies]
black = "^21.10b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
exclude = '''
/(
  | migrations
)/
'''
